fix: Windows Ctrl-C shutdown and comprehensive platform improvements
Major fixes for Windows signal handling and cross-platform compatibility:

Signal Handling & Thread Safety:
- Fixed server/client hanging on Ctrl-C on Windows by using C11 atomic operations
- Windows creates a new thread for SIGINT, requiring atomic_bool for proper cross-thread visibility
- Replaced all volatile bool shutdown flags with atomic_bool on all platforms
- Added atomic_load/atomic_store for all thread synchronization flags
- Implemented Windows-aware interruptible_usleep with proper Sleep() handling

Platform Abstraction Enhancements:
- Added comprehensive Windows POSIX compatibility macros (_CRT_SECURE_NO_WARNINGS, etc)
- Fixed socket type consistency (int → socket_t throughout codebase)
- Added proper socket shutdown sequence for Windows (shutdown() before close())
- Implemented aligned_alloc, clock_gettime, gmtime_r for Windows
- Added standard file descriptor constants (STDIN_FILENO, etc) for Windows
- Fixed terminal operations with proper file descriptor handling

Thread & Synchronization Improvements:
- Enhanced Windows thread implementation with detailed debug logging
- Fixed thread join operations with proper WaitForSingleObject handling
- Added comprehensive mutex/rwlock/condition variable debugging
- Improved socket initialization with INVALID_SOCKET_VALUE consistency

Terminal & Console Updates:
- Refactored terminal operations to use proper TTY file descriptors
- Added console_clear() wrapper for cross-platform screen clearing
- Fixed terminal_flush() to accept file descriptor parameter
- Enhanced Windows terminal implementation with proper escape sequence handling

Network & Socket Fixes:
- Fixed socket error handling with proper WSAGetLastError() on Windows
- Added socket_shutdown() for clean connection termination
- Improved recv timeout handling for Windows sockets
- Fixed POSIX socket includes and compatibility

SIMD & Performance:
- Cleaned up SIMD common headers and removed redundant includes
- Fixed output buffer memory management
- Improved ASCII frame generation with proper alignment

Testing & Reliability:
- All threads now properly exit on shutdown signal
- No more hanging processes or zombie threads
- Clean resource cleanup on both Windows and POSIX systems

This commit represents a major step forward in Windows support and
cross-platform reliability for ASCII-Chat.

+++ lib/common.c
@@ -351,7 +351,7 @@ void debug_memory_report(void) {
     // fprintf(stderr, "realloc calls: %zu\n", atomic_load(&g_mem.realloc_calls));
     fprintf(stderr, "free calls: %zu\n", free_calls);
     size_t diff = (malloc_calls + calloc_calls) - free_calls;
-    fprintf(stderr, "(malloc calls + calloc calls) - free calls = %zd\n", (ssize_t)diff);
+    fprintf(stderr, "(malloc calls + calloc calls) - free calls = %zu\n", diff);
 
     /* Only show allocations if tracking is enabled */
     if (g_mem.track_allocations && g_mem.head) {
